# students/k3339/Gneushev_Vladislav/lab3/docs/swagger.json
{
   "openapi":"3.1.0",
   "info":{
      "title":"FastAPI",
      "version":"0.1.0"
   },
   "paths":{
...
      "/api/assignments":{
         "get":{
            "tags":[
               "Назначения водителей на маршруты"
            ],
            "summary":"Получить список назначений",
            "operationId":"get_assigments_api_assignments_get",
            "parameters":[
               {
                  "name":"active",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "type":"boolean",
                     "description":"Активные назначения",
                     "title":"Active"
                  },
                  "description":"Активные назначения"
               },
               {
                  "name":"route_id",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "type":"integer",
                     "description":"ID маршрута",
                     "title":"Route Id"
                  },
                  "description":"ID маршрута"
               },
               {
                  "name":"ended_on_date",
                  "in":"query",
                  "required":false,
                  "schema":{
                     "type":"string",
                     "format":"date",
                     "description":"Дата завершения назначения",
                     "title":"Ended On Date"
                  },
                  "description":"Дата завершения назначения"
               }
            ],
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "type":"array",
                           "items":{
                              "$ref":"#/components/schemas/GetAssignmentSchema"
                           },
                           "title":"Response Get Assigments Api Assignments Get"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         },
         "post":{
            "tags":[
               "Назначения водителей на маршруты"
            ],
            "summary":"Добавить назначение",
            "operationId":"add_assignment_api_assignments_post",
            "parameters":[
               {
                  "name":"Authorization",
                  "in":"header",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Authorization"
                  }
               }
            ],
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/AddAssignmentSchema"
                     }
                  }
               }
            },
            "responses":{
               "201":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/GetAssignmentSchema"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/assignments/{assignment_id}":{
         "delete":{
            "tags":[
               "Назначения водителей на маршруты"
            ],
            "summary":"Удалить назначение",
            "operationId":"delete_assignment_api_assignments__assignment_id__delete",
            "parameters":[
               {
                  "name":"assignment_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "title":"Assignment Id"
                  }
               },
               {
                  "name":"Authorization",
                  "in":"header",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Authorization"
                  }
               }
            ],
            "responses":{
               "204":{
                  "description":"Successful Response"
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/assignments/{assignment_id}/end":{
         "post":{
            "tags":[
               "Назначения водителей на маршруты"
            ],
            "summary":"Завершить назначение",
            "operationId":"end_assignment_api_assignments__assignment_id__end_post",
            "parameters":[
               {
                  "name":"assignment_id",
                  "in":"path",
                  "required":true,
                  "schema":{
                     "type":"integer",
                     "title":"Assignment Id"
                  }
               },
               {
                  "name":"Authorization",
                  "in":"header",
                  "required":true,
                  "schema":{
                     "type":"string",
                     "title":"Authorization"
                  }
               }
            ],
            "requestBody":{
               "required":true,
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/EndAssignmentSchema"
                     }
                  }
               }
            },
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
...
      "/api/auth/login":{
         "post":{
            "tags":[
               "Авторизация"
            ],
            "summary":"Login",
            "operationId":"login_api_auth_login_post",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/LoginSchema"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/GetJWTTokenSchema"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/auth/register":{
         "post":{
            "tags":[
               "Авторизация"
            ],
            "summary":"Register",
            "operationId":"register_api_auth_register_post",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/RegisterSchema"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/GetJWTTokenSchema"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
      "/api/auth/refresh-token":{
         "post":{
            "tags":[
               "Авторизация"
            ],
            "summary":"Refresh Token",
            "operationId":"refresh_token_api_auth_refresh_token_post",
            "requestBody":{
               "content":{
                  "application/json":{
                     "schema":{
                        "$ref":"#/components/schemas/RefreshTokenSchema"
                     }
                  }
               },
               "required":true
            },
            "responses":{
               "200":{
                  "description":"Successful Response",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/GetJWTTokenSchema"
                        }
                     }
                  }
               },
               "422":{
                  "description":"Validation Error",
                  "content":{
                     "application/json":{
                        "schema":{
                           "$ref":"#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      }
   },
   "components":{
      "schemas":{
         "AddAssignmentSchema":{
            "properties":{
               "day_of_week":{
                  "$ref":"#/components/schemas/DayOfWeek"
               },
               "driver_id":{
                  "type":"integer",
                  "title":"Driver Id"
               },
               "route_id":{
                  "type":"integer",
                  "title":"Route Id"
               },
               "bus_id":{
                  "type":"integer",
                  "title":"Bus Id"
               }
            },
            "type":"object",
            "required":[
               "day_of_week",
               "driver_id",
               "route_id",
               "bus_id"
            ],
            "title":"AddAssignmentSchema"
         },
...
         "DayOfWeek":{
            "type":"string",
            "enum":[
               "monday",
               "tuesday",
               "wednesday",
               "thursday",
               "friday",
               "saturday",
               "sunday"
            ],
            "title":"DayOfWeek"
         },
...
         "EndAssignmentReason":{
            "type":"string",
            "enum":[
               "driver_ill",
               "bus_broken",
               "route_canceled",
               "bus_write_off",
               "other"
            ],
            "title":"EndAssignmentReason"
         },
         "EndAssignmentSchema":{
            "properties":{
               "reason":{
                  "$ref":"#/components/schemas/EndAssignmentReason"
               }
            },
            "type":"object",
            "required":[
               "reason"
            ],
            "title":"EndAssignmentSchema"
         },
         "GetAssignmentSchema":{
            "properties":{
               "id":{
                  "type":"integer",
                  "title":"Id"
               },
               "from_date":{
                  "type":"string",
                  "format":"date-time",
                  "title":"From Date"
               },
               "to_date":{
                  "anyOf":[
                     {
                        "type":"string",
                        "format":"date-time"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"To Date"
               },
               "end_reason":{
                  "anyOf":[
                     {
                        "$ref":"#/components/schemas/EndAssignmentReason"
                     },
                     {
                        "type":"null"
                     }
                  ]
               },
               "day_of_week":{
                  "$ref":"#/components/schemas/DayOfWeek"
               },
               "driver":{
                  "$ref":"#/components/schemas/GetDriverSchema"
               },
               "route":{
                  "$ref":"#/components/schemas/GetRouteSchema"
               },
               "bus":{
                  "$ref":"#/components/schemas/GetBusSchema"
               }
            },
            "type":"object",
            "required":[
               "id",
               "from_date",
               "to_date",
               "end_reason",
               "day_of_week",
               "driver",
               "route",
               "bus"
            ],
            "title":"GetAssignmentSchema"
         },
         "GetBusSchema":{
            "properties":{
               "id":{
                  "type":"integer",
                  "title":"Id"
               },
               "bus_type":{
                  "$ref":"#/components/schemas/GetBusTypeSchema"
               },
               "state_number":{
                  "type":"string",
                  "title":"State Number"
               }
            },
            "type":"object",
            "required":[
               "id",
               "bus_type",
               "state_number"
            ],
            "title":"GetBusSchema"
         },
         "GetBusTypeSchema":{
            "properties":{
               "id":{
                  "type":"integer",
                  "title":"Id"
               },
               "name":{
                  "type":"string",
                  "title":"Name"
               },
               "people_capacity":{
                  "type":"integer",
                  "title":"People Capacity"
               }
            },
            "type":"object",
            "required":[
               "id",
               "name",
               "people_capacity"
            ],
            "title":"GetBusTypeSchema"
         },
         "GetDriverClassSchema":{
            "properties":{
               "id":{
                  "type":"integer",
                  "title":"Id"
               },
               "name":{
                  "type":"string",
                  "title":"Name"
               }
            },
            "type":"object",
            "required":[
               "id",
               "name"
            ],
            "title":"GetDriverClassSchema"
         },
         "GetDriverSchema":{
            "properties":{
               "id":{
                  "type":"integer",
                  "title":"Id"
               },
               "first_name":{
                  "type":"string",
                  "title":"First Name"
               },
               "last_name":{
                  "type":"string",
                  "title":"Last Name"
               },
               "passport_info":{
                  "type":"string",
                  "title":"Passport Info"
               },
               "work_experience_months":{
                  "type":"integer",
                  "title":"Work Experience Months"
               },
               "driver_class":{
                  "$ref":"#/components/schemas/GetDriverClassSchema"
               },
               "salary_rub":{
                  "anyOf":[
                     {
                        "type":"integer"
                     },
                     {
                        "type":"null"
                     }
                  ],
                  "title":"Salary Rub"
               }
            },
            "type":"object",
            "required":[
               "id",
               "first_name",
               "last_name",
               "passport_info",
               "work_experience_months",
               "driver_class"
            ],
            "title":"GetDriverSchema"
         },
         "GetJWTTokenSchema":{
            "properties":{
               "token":{
                  "type":"string",
                  "title":"Token"
               },
               "expires_at":{
                  "type":"string",
                  "format":"date-time",
                  "title":"Expires At"
               }
            },
            "type":"object",
            "required":[
               "token",
               "expires_at"
            ],
            "title":"GetJWTTokenSchema"
         },
         "GetRouteSchema":{
            "properties":{
               "id":{
                  "type":"integer",
                  "title":"Id"
               },
               "name":{
                  "type":"string",
                  "title":"Name"
               },
               "start_point_name":{
                  "type":"string",
                  "title":"Start Point Name"
               },
               "end_point_name":{
                  "type":"string",
                  "title":"End Point Name"
               },
               "start_time":{
                  "type":"string",
                  "format":"time",
                  "title":"Start Time"
               },
               "end_time":{
                  "type":"string",
                  "format":"time",
                  "title":"End Time"
               },
               "interval_seconds":{
                  "type":"integer",
                  "title":"Interval Seconds"
               },
               "duration_seconds":{
                  "type":"integer",
                  "title":"Duration Seconds"
               },
               "length":{
                  "type":"integer",
                  "title":"Length"
               }
            },
            "type":"object",
            "required":[
               "id",
               "name",
               "start_point_name",
               "end_point_name",
               "start_time",
               "end_time",
               "interval_seconds",
               "duration_seconds",
               "length"
            ],
            "title":"GetRouteSchema"
         },
...
         "HTTPValidationError":{
            "properties":{
               "detail":{
                  "items":{
                     "$ref":"#/components/schemas/ValidationError"
                  },
                  "type":"array",
                  "title":"Detail"
               }
            },
            "type":"object",
            "title":"HTTPValidationError"
         },
         "LoginSchema":{
            "properties":{
               "username":{
                  "type":"string",
                  "title":"Username"
               },
               "password":{
                  "type":"string",
                  "title":"Password"
               }
            },
            "type":"object",
            "required":[
               "username",
               "password"
            ],
            "title":"LoginSchema"
         },
         "RefreshTokenSchema":{
            "properties":{
               "token":{
                  "type":"string",
                  "title":"Token"
               }
            },
            "type":"object",
            "required":[
               "token"
            ],
            "title":"RefreshTokenSchema"
         },
         "RegisterSchema":{
            "properties":{
               "username":{
                  "type":"string",
                  "title":"Username"
               },
               "password":{
                  "type":"string",
                  "title":"Password"
               }
            },
            "type":"object",
            "required":[
               "username",
               "password"
            ],
            "title":"RegisterSchema"
         },
...
         "ValidationError":{
            "properties":{
               "loc":{
                  "items":{
                     "anyOf":[
                        {
                           "type":"string"
                        },
                        {
                           "type":"integer"
                        }
                     ]
                  },
                  "type":"array",
                  "title":"Location"
               },
               "msg":{
                  "type":"string",
                  "title":"Message"
               },
               "type":{
                  "type":"string",
                  "title":"Error Type"
               }
            },
            "type":"object",
            "required":[
               "loc",
               "msg",
               "type"
            ],
            "title":"ValidationError"
         }
      }
   }
}